% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{brood_to_return}
\alias{brood_to_return}
\title{Convert a brood table to a return table}
\usage{
brood_to_return(bt)
}
\arguments{
\item{bt}{brood table with column names: Stock, BroodYear, and for each age (e.g., Age3). Must be in that format.}
}
\description{
Convert a brood table to a return table
}
