% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covariates_24}
\alias{covariates_24}
\title{Table of predictors}
\format{
\subsection{\code{covariates_24}}{

A tibble with 65 rows and 11 columns
\describe{
\item{ReturnYear}{Brood year}
\item{lag1_PDO lag2_PDO lag1_NPGO lag2_NPGO  lag1_fall_Nino3.4 lag2_fall_Nino3.4 smolt_sock lag1_log_socksmolt lag2_log_socksmolt pink_ind}{pssoble predictors: climate indices, sockeye smolt indices at Boneville Dam, and even-odd year index meants to represent the pink salmon cycle}
...
}
}
}
\usage{
covariates_24
}
\description{
Various predictors pulled from the internet
}
\keyword{datasets}
