% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mods.R
\name{fit_mods}
\alias{fit_mods}
\title{Prep data for each Stock/Age/n_years and  fit the models}
\usage{
fit_mods(
  dat,
  transformation = log,
  scale_x = FALSE,
  scale_y = FALSE,
  covariates = tibble(ReturnYear = numeric(0))
)
}
\arguments{
\item{dat}{tible returned from \code{setup_data} function}

\item{transformation}{transformation to be conducted on a response and predictors. default is log}

\item{scale_x}{boolean whether to scale the predictor data prior to fitting}

\item{scale_y}{boolean whether to scale the response data prior to fitting}

\item{covariates}{data frame of covariates to add to data. must contain "ReturnYear" field, which is used to perform the join with the salmon return data.}
}
\value{
tible
}
\description{
Prep data for each Stock/Age/n_years and  fit the models
}
\details{
Can take several seconds as it is conducting maximum likeihood optimization for many models (n mdel * n stocks * n forecasts)
}
