% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_fun.R
\name{forecast_fun}
\alias{forecast_fun}
\title{generate forecasts}
\usage{
forecast_fun(
  df = summer_chinook_2023,
  include = c("constIntOnly", "tvIntOnly", "tvIntSlope", "tvSlope", "constLM",
    "tvCRzeroInt", "constCRzeroInt", "tvInt"),
  perf_yrs = 15,
  wt_yrs = NULL
)
}
\arguments{
\item{df}{Input data. Either a return table or a brood table. Column names: Stock, ReturnYear (or BroodYear), and for each age (e.g., Age3). Must be in that format.}

\item{include}{vector of names of models to include. Default is all options}

\item{perf_yrs}{maximum number of years of predictions to include in performance metrics. Set to infinity to use a stretching window.}

\item{wt_yrs}{number of years of one-step-ahed predictions to use to calculate performance-based model weights (i.e., MAPE and RMSE). If null (the default) the same number of years is used to calculate weights as is used to evaluate performance of ensemble and individual models.}
}
\value{
A tibble with ensemble forecasts, forecasts from component models, and performance measures
}
\description{
generate forecasts
}
\details{
Model options are:
\itemize{
\item constLM --	Sibling regression with constant slope and intercept.
\item tvInt --	Sibling regression with  time-varying intercept.
\item tvSlope --	Sibling regression with time-varying slope.
\item tvIntSlope --	Sibling regression with time-varying slope and intercept.
\item tvCRzeroInt --	Time varying "cohort ratio" model. Time varying slope, Intercept=0.
\item constCRzeroInt --	Constant "cohort ratio" model. Constant slope, Intercept=0.
\item tvIntOnly --	Time-varying Intercept-only model. Random walk on return, no sibling predictor.
\item constIntOnly --	Constant Intercept-only model. Long-term average, no sibling predictor.
}
}
