% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_report.R
\name{forecast_report}
\alias{forecast_report}
\title{Render a report of forecast results}
\usage{
forecast_report(
  forecasts = NULL,
  mod_name = "MAPE_weight",
  stocks = NULL,
  forecast_year = NULL,
  df = NULL,
  include = NULL,
  perf_yrs = NULL,
  wt_yrs = NULL,
  output_file = "Forcast report.docx",
  output_dir = getwd()
)
}
\arguments{
\item{forecasts}{output of a call to \code{forecaste_fun}}

\item{mod_name}{name of the model that you want to plot forecasts for. options are:
"AICc_weight" ,"MAPE_weight","MeanSA_weight","RMSE_weight","constIntOnly","tvIntOnly","tvIntSlope","tvSlope","constLM","tvCRzeroInt","constCRzeroInt",or "tvInt"}

\item{forecast_year}{years to include in the table. default ot the last forecasts year in forecasts input}

\item{df}{Input data. Either a return table or a brood table. Column names: Stock, ReturnYear (or BroodYear), and for each age (e.g., Age3). Must be in that format.}

\item{include}{vector of names of models to include. Default is all options}

\item{perf_yrs}{maximum number of years of predictions to include in performance metrics. Set to infinity to use a stretching window.}

\item{wt_yrs}{number of years of one-step-ahed predictions to use to calculate performance-based model weights (i.e., MAPE and RMSE). If null (the default) the same number of years is used to calculate weights as is used to evaluate performance of ensemble and individual models.}

\item{output_dir}{}
}
\description{
Render a report of forecast results
}
