% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{get_AIC}
\alias{get_AIC}
\title{Calculate AICc for a dlm model object given the data and npar}
\usage{
get_AIC(y, mod, npar, mod_type = "dlm")
}
\arguments{
\item{y}{response data}

\item{mod}{dlm model model object}

\item{npar}{number of parameters}

\item{mod_type}{either "dlm" if one of the 8 models fitted with \code{dlm} package or "RTMB" if the penalized dlm fit with the \code{RTMB} package.}
}
\value{
real, AICc
}
\description{
Calculate AICc for a dlm model object given the data and npar
}
