% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{get_npar}
\alias{get_npar}
\title{Get number of estimated parameters from a built dlm model object}
\usage{
get_npar(build)
}
\arguments{
\item{build}{built dlm model object}
}
\value{
integer. number of parameters
}
\description{
Get number of estimated parameters from a built dlm model object
}
