% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{plot observations and forecasts by age}
\usage{
make_plot(forecasts, mod_name)
}
\arguments{
\item{forecasts}{output of a call to \code{forecaste_fun}}

\item{mod_name}{name of the model that you want to plot forecasts for. options are:
"AICc_weight" ,"MAPE_weight","MeanSA_weight","RMSE_weight","constIntOnly","tvIntOnly","tvIntSlope","tvSlope","constLM","tvCRzeroInt","constCRzeroInt",or "tvInt"}
}
\value{
a ggplot object
}
\description{
The black points are observations and the red points are the forecasts. Thick part of errorbar spans 50\% prediciton ainteval and thin part spans 90\% interval.
}
