% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Make results table}
\usage{
make_table(forecasts, mod_name, forecast_year = NULL)
}
\arguments{
\item{forecasts}{output of a call to \code{forecaste_fun}}

\item{mod_name}{name of the model that you want to plot forecasts for. options are:
"AICc_weight" ,"MAPE_weight","MeanSA_weight","RMSE_weight","constIntOnly","tvIntOnly","tvIntSlope","tvSlope","constLM","tvCRzeroInt","constCRzeroInt",or "tvInt"}

\item{forecast_year}{years to include in the table. default ot the last forecasts year in forecasts input}
}
\description{
Make results table
}
