% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_dlm.R
\name{pen_dlm}
\alias{pen_dlm}
\title{penalized DLM in RTMB}
\usage{
pen_dlm(
  dat,
  form = formula("y~x"),
  regu = c(0.05, 0.05),
  gamma_shape = 10,
  gamma_scale = 1
)
}
\arguments{
\item{dat}{dataframe with response and predictors}

\item{form}{formula specifying linear predictor}

\item{regu}{to avoid the standard deviations shrinking too small and running into numerical issues, the log of the standard deviation multiplied be these values are added to the log-likelihood (subtracted from the negative log-likelihood). Defaults are .05 and .05 for the penalties on the mean of the coefficients and the year-to-year variability.}

\item{gamma_shape}{shape paramter for the gamma prior on the expential distributions rate paramters.}

\item{gamma_scale}{scale paramter for the gamma prior on the expential distributions rate paramters.}
}
\value{
a list with two components: the fitted TMB model object, which has the NLL and the linear predictors in the report(), as well as the outfut from the call to TMBhelper::fit_tmb(), which is used to optimize the model.
}
\description{
penalized DLM in RTMB
}
