% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_weights.R
\name{performance_weights}
\alias{performance_weights}
\title{Calculated weighted average forecasts}
\usage{
performance_weights(
  fits,
  perf_yrs = 15,
  wt_yrs = NULL,
  transformation = log,
  inverse_transformation = exp,
  scale_y = FALSE
)
}
\arguments{
\item{fits}{data frame returned by \code{fit_mods} function}

\item{perf_yrs}{maximum number of years of predictions to include in performance metrics. Set to infinity to use a stretching window.}

\item{wt_yrs}{number of years of one-step-ahed predictions to use to calculate performance-based model weights (i.e., MAPE and RMSE). If null (the default) the same number of years is used to calculate weights as is used to evaluate performance of ensemble and individual models.}

\item{transformation}{the trnasformation that was conducted on the response prior to fitting(default is log)}

\item{inverse_transformation}{the inverse of the transformation that was conducted on the response prior to fitting (default is exp)}

\item{scale_y}{boolean whether the response was scaled prior to fitting}
}
\value{
a data frame with predictions from component models and weighted average models based on AICc, MAPE, and RMSE weights.
}
\description{
Calculated weighted average forecasts
}
\details{
AICc weights are calculated as:
$exp(-.5 deltaAICc) / sum(exp(-.5 deltaAICc))$
MAPE and RMSE weights are calculated as, for example:
MAPE^-1 / sum(MAPE^-1)
\itemize{
\item \emph{model_name} is the name of the model used for forecast. It includes the compenent of the ensembles and the ensembles. Additionally, "best" is the model with the lowest MAPE over the perforance-evaluation period for a given year. Thgis is included to be be able to see the performance of picking the best-performaning model, whether than be an ensemble or one of the components, over multiple years. The "Model_name2" field indicates which model was the "best" for that year.
\item \emph{n_wts} = the number of years of forecasts and observations that were used to generate MAPE and RMSE weights for ensemble forecasts.
\item \emph{error} = error messages for models that failed to fit.
\item Individual year metrics:
\itemize{
\item \emph{Er} = forecast error
\item \emph{AEr} = absolute forecast error
\item \emph{PE}= percent errr,
\item \emph{APE} = absolute percent error
\item \emph{SQE} = square error
\item \emph{logQ} = log (forecast/observed)- AlogQ = absolute logQ
}
\item Multi-year bias metrics
\itemize{
\item \emph{MEr} = mean error
\item \emph{MPE} = mean percent error
\item \emph{MeanlogQ} = mean logQ
}
\item Multi-year precision metrics
\itemize{
\item \emph{RMSE} = root mean square error
\item \emph{MAer} = mean absolute error
\item \emph{MAPE} = mean absolute percent error
\item \emph{MeanSA} = mean absolute logQ
}
\item \emph{log_sd} = standard deviation from observed in log space for a the last \emph{n_years} of forecasts (i.e., (sqrt(sum(logQ^2)/n)))
\item \emph{n_sd} = the number of years that were used to calculate sample standard deviation from observed
}
}
