% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{return_to_brood}
\alias{return_to_brood}
\title{Convert a Return Table to a Brood Table}
\usage{
return_to_brood(rt, by_stock = TRUE)
}
\arguments{
\item{rt}{return table with column names: Stock, ReturnYear, and for each age (e.g., Age3). Must be in that format.}

\item{by_stock}{boolean for whether the data contians a Stock column.}
}
\description{
Convert a Return Table to a Brood Table
}
