% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_data.R
\name{setup_data}
\alias{setup_data}
\title{Setup data}
\usage{
setup_data(df, mod_list = mod_funs(), n_forecasts = 20)
}
\arguments{
\item{df}{Input data. Either a return table or a brood table. Column names: Stock, ReturnYear (or BroodYear), and for each age (e.g., Age3). Must be in that format.}

\item{mod_list}{list of model functions as returned from \code{mod_funs} function call}

\item{n_forecasts}{number of one-year ahead forecasts to conduct}
}
\value{
tibble with list columns
}
\description{
puts data into a list column with different subsets (different numbers of years excluded) in each row, for fitting
}
