% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{stretching_mean}
\alias{stretching_mean}
\title{Stretching mean}
\usage{
stretching_mean(x, window_size = Inf)
}
\arguments{
\item{x}{a vector of values}

\item{window_size}{the maximum number of values to include in mean}
}
\description{
Stretching mean
}
