% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{stretching_samp_sd}
\alias{stretching_samp_sd}
\title{Stretching sample SD}
\usage{
stretching_samp_sd(x, window_size = Inf, sample_sd = FALSE)
}
\arguments{
\item{x}{a vector of errors}

\item{window_size}{the maximum number of values to include in mean}

\item{sample_sd}{boolean whether to subtract one from denominator so as to calculate a sample standard deviation. I think this would only be approporiate if bias correction were being applied.}
}
\description{
Stretching sample SD
}
\examples{
vec <- 1:10
stretching_samp_sd(vec, window_size = 5)
}
