% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summer_chinook_2024}
\alias{summer_chinook_2024}
\title{Summer chinook returns through 2024}
\format{
\subsection{\code{summer_chinook_2024}}{

A tibble with 28 rows and 6 columns
\describe{
\item{BroodYear}{Brood year}
\item{Age3, Age4, Age5, Age6}{returns by age}
\item{Stock}{name of stock}
...
}
}
}
\usage{
summer_chinook_2024
}
\description{
Output of TAC summer Chinook managment season run reconstructions through return to the Columbia River mouth year 2024
}
\keyword{datasets}
