% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mods.R
\name{fit_mods}
\alias{fit_mods}
\title{Prep data for each Stock/Age/n_years and  fit the models}
\usage{
fit_mods(dat)
}
\arguments{
\item{dat}{tible returned from \code{setup_data} function}
}
\value{
tible
}
\description{
Prep data for each Stock/Age/n_years and  fit the models
}
\details{
Can take several seconds as it is conducting maximum likeihood optimization for many models (n mdel * n stocks * n forecasts)
}
