% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_fun.R
\name{forecast_fun}
\alias{forecast_fun}
\title{generate forecasts}
\usage{
forecast_fun(
  df = summer_chinook_2023,
  type = "brood",
  include = c("constIntOnly", "tvIntOnly", "tvIntSlope", "tvSlope", "constLM",
    "tvCRzeroInt", "constCRzeroInt", "tvInt"),
  perf_yrs = 15
)
}
\arguments{
\item{df}{Input data. Either a return table or a brood table. Column names: Stock, ReturnYear (or BroodYear), and for each age (e.g., Age3). Must be in that format.}

\item{type}{whether df is a "return" or "brood" table. Should be one of those values.}

\item{include}{vector of names of models to include. Default is all options}

\item{perf_yrs}{maximum number of years of predictions to include in performance metrics. Set to infinity to use a stretching window.}
}
\value{
A tibble with ensemble forecasts, forecasts from component models, and performance measures
}
\description{
generate forecasts
}
\details{
Model options are:
\itemize{
\item constLM --	Sibling regression with constant slope and intercept.
\item tvInt --	Sibling regression with  time-varying intercept.
\item tvSlope --	Sibling regression with time-varying slope.
\item tvIntSlope --	Sibling regression with time-varying slope and intercept.
\item tvCRzeroInt --	Time varying "cohort ratio" model. Time varying slope, Intercept=0.
\item constCRzeroInt --	Constant "cohort ratio" model. Constant slope, Intercept=0.
\item tvIntOnly --	Time-varying Intercept-only model. Random walk on return, no sibling predictor.
\item constIntOnly --	Constant Intercept-only model. Long-term average, no sibling predictor.
}
}
