% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_fun.R
\name{forecast_fun}
\alias{forecast_fun}
\title{generate forecasts}
\usage{
forecast_fun(
  df = summer_chinook_2024,
  include = c("constIntOnly", "tvIntOnly", "tvSlope", "constLM", "tvCRzeroInt",
    "constCRzeroInt", "tvInt"),
  transformation = log,
  inverse_transformation = exp,
  scale_x = FALSE,
  scale_y = FALSE,
  perf_yrs = 15,
  wt_yrs = NULL,
  covariates = tibble(ReturnYear = numeric(0)),
  include_youngest = FALSE,
  ...
)
}
\arguments{
\item{df}{Input data. A return table oColumn names: Stock, ReturnYear, and abundance for each age (e.g., Age3). Must be in that format.}

\item{include}{vector of names of models to include. Default is all options}

\item{transformation}{transformation to be conducted on a response and predictors. default is log}

\item{inverse_transformation}{the inverse of the transformation that was conducted on the response prior to fitting (default is exp)}

\item{scale_x}{boolean whether to scale the predictor data prior to fitting}

\item{scale_y}{boolean whether to scale the response data prior to fitting}

\item{perf_yrs}{maximum number of years of predictions to include in performance metrics. Set to infinity to use a stretching window.}

\item{wt_yrs}{number of years of one-step-ahed predictions to use to calculate performance-based model weights (i.e., MAPE and RMSE). If null (the default) the same number of years is used to calculate weights as is used to evaluate performance of ensemble and individual models.}

\item{covariates}{data frame of covariates to add to data. must contain "ReturnYear" or "BroodYear" field, which is used to perform the join with the salmon return data. Missing values in covariates are ok, and are fit as random effects with normal(0,1) hyperdistribution. Therefore it is importand to scale covariates especially if there are missing values.}

\item{...}{other arguments from \code{mod_funs} or \code{performance_weights}}
}
\value{
A list of two dataframes: \emph{fits} is the output of the call to \code{fit_mods()} and potentially useful information about models that failed to fit. \emph{forecasts} is the output of a call to \code{performance_weights()} with ensemble forecasts, forecasts from component models, and performance measures
}
\description{
generate forecasts
}
\details{
The models available in this package are variations of a sibling regression model with time-varying intercept and slope:
\deqn{\left\{\begin{aligned}\mathrm{log}(y_{a,t}) & =\alpha_t + \mathrm{log}(y_{a-1,t-1}) \beta_t + v_t, & \quad &v_t \sim \mathcal{N}\left(0, V_t\right), \\\alpha_t & = \alpha_{t-1} + w_{\alpha, t}, & \quad & w_{\alpha, t} \sim \mathcal{N}\left(0, W_{\alpha, t}\right), \\\beta_t & = \beta_{t-1} + w_{\beta, t}, & \quad & w_{\beta, t} \sim \mathcal{N}\left(0, W_{\beta, t}\right),\end{aligned}\right.}

where \eqn{\alpha_t} and \eqn{\beta_t} are an intercept and a slope for returns of the previous age in the previous year, respectively, and both are allowed to vary across years as random walks with process error variances \eqn{w_{\alpha, t}} and \eqn{w_{\beta, t}} respectively. The residual \eqn{v_t} is assumed to be normally distributed around zero with variance \eqn{V_t}. Together with a vague prior distribution for the values of \eqn{\alpha_0} and \eqn{\beta_0} (the slope and intercept prior to the first year) these equations define the "full" sibling regression models. However, this "full" model may be overly complex for optimal prediction, and is difficult to fit in practice. Simplified versions of this model are included, as described below.

Model options are:
\itemize{
\item constLM --	Sibling regression with constant slope and intercept.
\item tvInt --	Sibling regression with  time-varying intercept.
\item tvSlope --	Sibling regression with time-varying slope.
\item tvIntSlope --	Sibling regression with time-varying slope and intercept (i.e., the "full" mode). This model is not included by default because it generally converges to one of the simpler model. In other words, it is exceedingly rare for the data to support this complex of a model and this including it in an ensemble leads to double weighting one of the simpler models.
\item tvCRzeroInt --	Time varying "cohort ratio" model. Time varying slope, Intercept=0.
\item constCRzeroInt --	Constant "cohort ratio" model. Constant slope, Intercept=0.
\item tvIntOnly --	Time-varying Intercept-only model. Random walk on return, no sibling predictor.
\item constIntOnly --	Constant Intercept-only model. Long-term average, no sibling predictor.
\item PenDlm -- A dynamic linear model the like "tvIntSlope" but with penalized-complexity priors on the year-to-year variation and the mean of the covariate values. This model also can accommodate additional predictors (e.g., environmental covariates). This model is not included by defaults
}
}
