% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funs.R
\name{get_AIC}
\alias{get_AIC}
\title{Calculate AICc for a dlm model object given the data and npar}
\usage{
get_AIC(y, mod, npar)
}
\arguments{
\item{y}{response data}

\item{mod}{dlm model model object}

\item{npar}{number of parameters}
}
\value{
real, AICc
}
\description{
Calculate AICc for a dlm model object given the data and npar
}
