% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_funs.R
\name{mod_funs}
\alias{mod_funs}
\title{functions to build models}
\usage{
mod_funs(
  include = c("constIntOnly", "tvIntOnly", "tvIntSlope", "tvSlope", "constLM",
    "tvCRzeroInt", "constCRzeroInt", "tvInt"),
  penDLM_formula = NULL,
  penDLM_regu = NULL,
  penDLM_gamma_shape = NULL,
  penDLM_gamma_scale = NULL
)
}
\arguments{
\item{include}{vector of names of models to include. Default is all options}

\item{penDLM_formula}{an optional r formula object to pass to the penalized DLM model. Default is NULL, which would use the defaulte value formula("y~x") from the pen_dlm function.}

\item{penDLM_regu}{scaler for the addition of the log of standard deviations to the log-likelihood as a regularizer. See penealized_dlm help for more details}

\item{penDLM_gamma_shape}{shape parameter for the gamma prior on the exponential distribution rate parameter. See penealized_dlm help for more details}

\item{penDLM_gamma_scale}{scale parameter for the gamma prior on the exponential distribution rate parameter.See penealized_dlm help for more details}
}
\value{
a list of functions that return (unoptimized) dlm model objects
}
\description{
functions to build models
}
\details{
Model options are:
\itemize{
\item constLM --	Sibling regression with constant slope and intercept.
\item tvInt --	Sibling regression with  time-varying intercept.
\item tvSlope --	Sibling regression with time-varying slope.
\item tvIntSlope --	Sibling regression with time-varying slope and intercept.
\item tvCRzeroInt --	Time varying "cohort ratio" model. Time varying slope, Intercept=0.
\item constCRzeroInt --	Constant "cohort ratio" model. Constant slope, Intercept=0.
\item tvIntOnly --	Time-varying Intercept-only model. Random walk on return, no sibling predictor.
\item constIntOnly --	Constant Intercept-only model. Long-term average, no sibling predictor.
}
}
