% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_dlm.R
\name{pen_dlm}
\alias{pen_dlm}
\title{penalized DLM in RTMB}
\usage{
pen_dlm(dat, form = formula("y~x"), exp_pen = c(5, 5), regu = c(0.01, 0.01))
}
\arguments{
\item{dat}{dataframe with response and predictors}

\item{form}{formula specifying linear predictor}

\item{exp_pen}{rate parameters for the exponential penalties on the standard deviation of the mean of the paramemets and the random walk innovations, respectively. Default is 5 and 5}

\item{regu}{to avoid the standard deviations shrinking too small and running into numerical issues, the log of the standard deviation multiplied be these values are added to the log-likelihood (subtracted from the negative log-likelihood). Defaults are .01 and .01 for the two penalties.}
}
\value{
a list with two components: the fitted TMB model object, which has the NLL and the linear predictors in the report(), as well as the outfut from the call to TMBhelper::fit_tmb(), which is used to optimize the model.
}
\description{
penalized DLM in RTMB
}
