% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalized_dlm.R
\name{pen_dlm}
\alias{pen_dlm}
\title{penalized DLM in RTMB}
\usage{
pen_dlm(
  dat,
  form = stats::formula("y~x"),
  regu = c(0.01, 0.01),
  gamma_shape = 10,
  gamma_scale = 1,
  exp_rate = c(1, 1)
)
}
\arguments{
\item{dat}{dataframe with response and predictors}

\item{form}{formula specifying linear predictor}

\item{regu}{to avoid the standard deviations shrinking too small and running into numerical issues, the log of the standard deviation multiplied be these values are added to the log-likelihood (subtracted from the negative log-likelihood). Defaults are .05 and .05 for the penalties on the mean of the coefficients and the year-to-year variability.}

\item{gamma_shape}{shape paramter for the gamma prior on the expential distributions rate paramters.}

\item{gamma_scale}{scale paramter for the gamma prior on the expential distributions rate paramters.}
}
\value{
a list with two components: the fitted TMB model object, which has the NLL and the linear predictors in the report(), as well as the outfut from the call to TMBhelper::fit_tmb(), which is used to optimize the model.
}
\description{
function to fit a penalized dynamic linear regression model.

The penalized complexity model puts penalties on the across-year means \eqn{\bar{\beta}} of each coefficient \eqn{\beta_t} for each year \eqn{t}, and the standard deviation of the steps in the random walk. So if the coefficients are:

\deqn{\beta_t = \beta_{t-1} + \omega_t \\ \omega_t \sim \mathcal{N}(0,\sigma)}

\deqn{\bar{\beta}\sim \mathcal{N}(0,\tau)}

This model puts exponential-gamma penalties on all $\tau$ and $\sigma$ parameters, for which there is a unique parameter for each predictor in the model:

\deqn{\tau,~\sigma \sim \text{exp}(\lambda) \\ \lambda \sim \text{Gamma}(\text{Shape}=10,~\text{Scale}=1)}

where two unique \eqn{\lambda} parameters are fit for every predictor in the model.

Additionally, \eqn{0.05*\text{log}(\tau, \sigma)} for all \eqn{\tau}s and \eqn{\sigma}s is added to the log-likelihood to keep their values from shrinking so small as to cause numerical problems during optimization.
}
