% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_weights.R
\name{performance_weights}
\alias{performance_weights}
\title{Calculated weighted average preds}
\usage{
performance_weights(fits)
}
\arguments{
\item{fits}{data frame returned by \code{fit_mods} function}
}
\value{
a data frame with predicitons from component models and weighted average models based on AICc, MAPE, and RMSE weights.
}
\description{
Calculated weighted average preds
}
\details{
AICc weights are calculated as exp(-.5\emph{deltaAICc) / sum(exp(-.5}deltaAICc))
MAPE and RMSE weights are calculated as, for example, Mape^-1 / sum(mape^-1)
}
