% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_data.R
\name{setup_data}
\alias{setup_data}
\title{Setup data}
\usage{
setup_data(df, type = "return", n_forecasts = 20)
}
\arguments{
\item{df}{Input data. Either a return table or a brood table. Column names: Stock, ReturnYear (or BroodYear), and for each age (e.g., Age3). Must be in that format.}

\item{type}{whether df is a "return" or "brood" table. Should be one of those values.}

\item{n_eval}{number of one year-ahead predictions to make (i.e., maximum number of years to exclude)}
}
\value{
tibble with list columns
}
\description{
puts data into a list column with different subsets (different numbers of years excluded) in each row, for fitting
}
